+++ src/exception/QcacheException.php
@@ -4,10 +4,10 @@ namespace acet\qcache\exception;
 
 use Exception;
 
-class QCacheException extends Exception
+class QcacheException extends Exception
 {
     public function __construct($message='')
     {
-        parent::__construct("QCache: $message");
+        parent::__construct("Qcache: $message");
     }
 }